requirements.txt
pyserial
psutil
PySensors
nanopb